Testing changes

+++ grocerystore/testing/test_admin.py
@@ -12,7 +12,7 @@ test_all_products_unauthorized_401: Except Admin. No Permission  : 401
 
 
 def test_all_products_200(client, admin_token_header):
-    response = client.get('/admin/get_items', headers=admin_token_header)
+    response = client.get('/admin/get_items', headers={"Authorization": admin_token_header})
     assert response.status_code == 200
     assert "Items Fetched Successfully"
 
@@ -42,7 +42,7 @@ def test_add_product_201(client, admin_token_header):
         "price": 100,
         "quantity": 100
     }]
-    response = client.post('/admin/create_items', json.dumps(data), headers=admin_token_header)
+    response = client.post('/admin/create_items', json.dumps(data), headers={"Authorization": admin_token_header})
     assert response.status_code == 201
     assert response.json()["DB Status"] == "Item Added Successfully"
 
@@ -81,7 +81,7 @@ def test_update_product_200(client, admin_token_header):
         "price": 200,
         "quantity": 200
     }
-    response = client.put('admin/update_item/10', json.dumps(data), headers=admin_token_header)
+    response = client.put('admin/update_item/4', json.dumps(data), headers={"Authorization": admin_token_header})
     assert response.status_code == 200
     assert "Items Updated Successfully"
 
@@ -94,7 +94,7 @@ def test_update_product_unauthorized_401(client):
         "price": 200,
         "quantity": 200
     }
-    response = client.put('admin/update_item/10', json.dumps(data), headers={"Authorization": "Bearer fake_token"})
+    response = client.put('admin/update_item/4', json.dumps(data), headers={"Authorization": "Bearer fake_token"})
     assert response.status_code == 401
     assert "UnAuthorized User"
 
@@ -107,7 +107,7 @@ def test_update_product_not_found_404(client, admin_token_header):
         "price": 200,
         "quantity": 200
     }
-    response = client.put('/update_item/100', json.dumps(data), headers=admin_token_header)
+    response = client.put('/update_item/100', json.dumps(data), headers={"Authorization": admin_token_header})
     assert response.status_code == 404
     assert "Item Not Found"
 
@@ -120,7 +120,7 @@ def test_update_product_no_changes_302(client, admin_token_header):
         "price": 200,
         "quantity": 200
     }
-    response = client.put('admin/update_item/10', json.dumps(data), headers=admin_token_header)
+    response = client.put('admin/update_item/4', json.dumps(data), headers={"Authorization": admin_token_header})
     assert response.status_code == 302
     assert "No Changes Detected"
 
@@ -138,19 +138,19 @@ test_delete_product_not_found_404: Item ID entered, Not Found    : 404
 
 
 def test_delete_product_200(client, admin_token_header):
-    response = client.delete('/admin/delete_item/2', headers=admin_token_header)
+    response = client.delete('/admin/delete_item/4', headers={"Authorization": admin_token_header})
     assert response.status_code == 200
     assert "Item Deleted Successfully"
 
 
 def test_delete_product_unauthorized_401(client):
-    response = client.delete('/admin/delete_item/2', headers={"Authorization": "Bearer fake_token"})
+    response = client.delete('/admin/delete_item/4', headers={"Authorization": "Bearer fake_token"})
     assert response.status_code == 401
     assert "UnAuthorized User"
 
 
 def test_delete_product_not_found_404(client, admin_token_header):
-    response = client.delete('/admin/delete_item/100', headers=admin_token_header)
+    response = client.delete('/admin/delete_item/100', headers={"Authorization": admin_token_header})
     assert response.status_code == 404
     assert "Item Not Found"
 
@@ -168,7 +168,7 @@ test_delete_product_not_found_404: Item ID entered, Not Found    : 404
 
 
 def test_view_orders_200(client, admin_token_header):
-    response = client.get('/admin/view_orders', headers=admin_token_header)
+    response = client.get('/admin/view_orders', headers={"Authorization": admin_token_header})
     assert response.status_code == 200
     assert "Items Fetched Successfully"
 
@@ -197,7 +197,7 @@ test_add_discount_coupon_invalid_date_401: Invalid Date Entered  : 401
 #         "discount_percentage": 10,
 #         "valid_till": 2030-12-12
 #     }]
-#     response = client.post('/admin/add_discount_coupon', json.dumps(data), headers=admin_token_header)
+#     response = client.post('/admin/add_discount_coupon', json.dumps(data), headers={"Authorization": admin_token_header})
 #     assert response.status_code == 200
 #     assert response.json()['status'] == 'Coupons Added Successfully'
 
@@ -215,11 +215,11 @@ def test_add_discount_coupon_not_authorized_401(client):
 
 def test_add_discount_coupon_invalid_date_401(client, admin_token_header):
     data = [{
-        "coupon_code": "CoDe",
+        "coupon_code": "Grocery",
         "discount_percentage": 10,
         "valid_till": 2022-10-10
     }]
-    response = client.post('/admin/add_discount_coupon', json.dumps(data), headers=admin_token_header)
+    response = client.post('/admin/add_discount_coupon', json.dumps(data), headers={"Authorization": admin_token_header})
     assert response.status_code == 401
     assert "Invalid Date Entered"
 
@@ -236,7 +236,7 @@ test_show_discount_coupon_not_authorized_401: No User Permission : 401
 
 
 def test_show_discount_coupon_200(client, admin_token_header):
-    response = client.get('/admin/show_discount_coupon', headers=admin_token_header)
+    response = client.get('/admin/show_discount_coupon', headers={"Authorization": admin_token_header})
     assert response.status_code == 200
     assert "Discount Coupons Fetched Successfully"
